Fix Diplomat/Spy actions against a victim on the same tile as the Diplomat
The offending line cause silent failure unless the Spy/Diplomat action
dialog of another unit is open. Remove it in stead of inverting it. No
matter if the check is there or not the Spy/Diplomat action dialog currently
showing will remain the same. Without the check the new order will be queued
in stead of silently forgotten.

See bug #21595

git-svn-id: svn://svn.gna.org/svn/freeciv/branches/S2_5@24364 a0f10bec-cc02-0410-94fc-a9cfff90b4cd