Pass the acting diplomat from the generic to the specific client code
All packets asking follow up questions about a spy action contain the ID of
the acting diplomat. The generic client code already use it.

Change the interface between the generic and the specific client code so the
diplomat is forwarded to the client specific code. Having the real diplomat
as a parameter is useful when fixing a bug caused by a client specific way
of guessing what diplomat a follow up question refers to.

See patch #4502

git-svn-id: svn://svn.gna.org/svn/freeciv/branches/S2_5@24484 a0f10bec-cc02-0410-94fc-a9cfff90b4cd