Also search multiarch folders when looking for Qt 5 (and using GCC)
Some systems, like recent versions of Debian, store some architecture specific
headers and libraries in architecture specific locations. This makes it possible
to install a foreign version of a library for your cross compiler without
removing the version built for your own architecture. The architecture specific
location includes the multiarch tuple. The multiarch tuple is a GNU triplet that
may be normalized or have extensions.

Unless the location of Qt 5 libraries and Qt 5 headers are given on the command
line the ./configure script will search for them. It has one list of locations
to look for headers and another to look for libraries.

This change tries to find the multiarch tuple. If found multiarch locations are
added to the list of locations to be searched. If not found only non multiarch
locations are searched.

If the selected C compiler is a recent version of GCC its --print-multiarch
option will be used to get the multiarch tuple. Since it comes from the C
compiler it should return the correct multiarch tuple when cross compiling.

Since the GNU triplet in many cases is identical to the multiarch tuple it is
used when the selected C compiler don't supports the --print-multiarch option.
It is created from the AC_CANONICAL_HOST variables $host_cpu and $host_os. This
trick was stolen from a patch written by Marko Lindqvist <cazfi>

Reviewed by Emmet Hikory <persia>

See bug #22204

git-svn-id: svn://svn.gna.org/svn/freeciv/branches/S2_5@25364 a0f10bec-cc02-0410-94fc-a9cfff90b4cd