A transport assigned to war isn't (yet) on a spy mission.
Use AIUNIT_NONE in stead of AIUNIT_ATTACK to signal that a transport unit
that also is an attack unit is to be used for attacking. This avoids a crash
when the unit capable of attacking and transporting also is capable of doing
spy actions to cities.

Reported by Marko Lindqvist <cazfi>

See bug #23107

git-svn-id: svn://svn.gna.org/svn/freeciv/branches/S2_5@27535 a0f10bec-cc02-0410-94fc-a9cfff90b4cd